fix patch

+++ kubernetes/apps/backup/backrest/app/resources/init.sh
@@ -4,7 +4,9 @@ set -e
 CONFIG_PATH=/app/config.json
 TEMP_CONFIG_PATH=/tmp/config.json.tmp
 
-echo "{\"repos\": []}" > $CONFIG_PATH
+if [ ! -f "$CONFIG_PATH" ]; then
+  echo "{\"repos\": []}" > $CONFIG_PATH
+fi
 cat $CONFIG_PATH | jq
 jq ".instance = \"${CLUSTER_CNAME}\"" $CONFIG_PATH > $TEMP_CONFIG_PATH && cp $TEMP_CONFIG_PATH $CONFIG_PATH
 jq ".version = 4" $CONFIG_PATH > $TEMP_CONFIG_PATH && cp $TEMP_CONFIG_PATH $CONFIG_PATH
